hoist StorybookDecorator into it's own component
