+++ packages/studio-base/src/components/AppBar/AddPanel.stories.tsx
@@ -2,13 +2,13 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { AddPanelMenu } from "@foxglove/studio-base/components/AppBar/AddPanelMenu";
-import { Wrapper } from "@foxglove/studio-base/components/AppBar/index.stories";
+import { AddPanelMenu } from "./AddPanelMenu";
+import { StorybookDecorator } from "./StorybookDecorator";
 
 export default {
   title: "components/AppBar/AddPanelMenu",
   component: AddPanelMenu,
-  decorators: [Wrapper],
+  decorators: [StorybookDecorator],
 };
 
 export function Default(): JSX.Element {
