+++ packages/studio-base/src/components/AppBar/StorybookDecorator.tsx
@@ -1,0 +1,72 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import { Story } from "@storybook/react";
+import { DndProvider } from "react-dnd";
+import { HTML5Backend } from "react-dnd-html5-backend";
+
+import MockMessagePipelineProvider from "@foxglove/studio-base/components/MessagePipeline/MockMessagePipelineProvider";
+import Panel from "@foxglove/studio-base/components/Panel";
+import PanelCatalogContext, {
+  PanelCatalog,
+  PanelInfo,
+} from "@foxglove/studio-base/context/PanelCatalogContext";
+import MockCurrentLayoutProvider from "@foxglove/studio-base/providers/CurrentLayoutProvider/MockCurrentLayoutProvider";
+
+const SamplePanel1 = function () {
+  return <div></div>;
+};
+SamplePanel1.panelType = "sample";
+SamplePanel1.defaultConfig = {};
+
+const SamplePanel2 = function () {
+  return <div></div>;
+};
+SamplePanel2.panelType = "sample2";
+SamplePanel2.defaultConfig = {};
+
+const MockPanel1 = Panel(SamplePanel1);
+const MockPanel2 = Panel(SamplePanel2);
+
+const allPanels: PanelInfo[] = [
+  { title: "Regular Panel BBB", type: "Sample1", module: async () => ({ default: MockPanel1 }) },
+  { title: "Regular Panel AAA", type: "Sample2", module: async () => ({ default: MockPanel2 }) },
+
+  {
+    title: "Preconfigured Panel AAA",
+    type: "Sample1",
+    description: "Panel description",
+    module: async () => ({ default: MockPanel1 }),
+    config: { text: "def" },
+  },
+  {
+    title: "Preconfigured Panel BBB",
+    type: "Sample2",
+    module: async () => ({ default: MockPanel1 }),
+    config: { num: 456 },
+  },
+];
+
+class MockPanelCatalog implements PanelCatalog {
+  public getPanels(): readonly PanelInfo[] {
+    return allPanels;
+  }
+  public getPanelByType(type: string): PanelInfo | undefined {
+    return allPanels.find((panel) => !panel.config && panel.type === type);
+  }
+}
+
+export function StorybookDecorator(StoryFn: Story): JSX.Element {
+  return (
+    <DndProvider backend={HTML5Backend}>
+      <PanelCatalogContext.Provider value={new MockPanelCatalog()}>
+        <MockCurrentLayoutProvider>
+          <MockMessagePipelineProvider>
+            <StoryFn />
+          </MockMessagePipelineProvider>
+        </MockCurrentLayoutProvider>
+      </PanelCatalogContext.Provider>
+    </DndProvider>
+  );
+}
